Refactor results from JobsAccordion into its own component.
