+++ app/src/scripts/dataset/run/index.js
@@ -3,12 +3,12 @@
 
 import React from 'react'
 import PropTypes from 'prop-types'
-import actions from './dataset.actions'
-import WarnButton from '../common/forms/warn-button.jsx'
+import actions from '../dataset.actions'
+import WarnButton from '../../common/forms/warn-button.jsx'
 import moment from 'moment'
-import FileTree from '../common/partials/file-tree.jsx'
+import Results from './results.jsx'
 import { Accordion, Panel } from 'react-bootstrap'
-import markdown from '../utils/markdown'
+import markdown from '../../utils/markdown'
 
 class JobAccordion extends React.Component {
   // life cycle methods ------------------------------------------------------------
@@ -27,7 +27,12 @@ class JobAccordion extends React.Component {
           <span className="inner">
             {this._support(run)}
             {this._parameters(run)}
-            {this._results(run, 'results')}
+            <Results
+              run={run}
+              acknowledgements={this.props.acknowledgements}
+              displayFile={this.props.displayFile}
+              toggleFolder={this.props.toggleFolder}
+            />
             {this._logs(run)}
             {this._batchStatus(run)}
           </span>
@@ -84,67 +89,6 @@ class JobAccordion extends React.Component {
         {this._canceledMesssage(run)}
       </div>
     )
-  }
-
-  _results(run, type) {
-    if (run[type] && run[type].length > 0) {
-      return (
-        <Accordion accordion className="results">
-          <Panel
-            className="fade-in"
-            header={type}
-            key={run._id}
-            eventKey={run._id}>
-            <div className="app-acknowledgements">
-              <label>Acknowledgements</label>
-              <div
-                className="markdown"
-                dangerouslySetInnerHTML={markdown.format(
-                  this.props.acknowledgements,
-                )}
-              />
-            </div>
-            <hr />
-            <span className="download-all">
-              <WarnButton
-                icon="fa-download"
-                message=" DOWNLOAD All"
-                prepDownload={actions.getResultDownloadTicket.bind(
-                  this,
-                  run.snapshotId,
-                  run._id,
-                  { path: 'all-' + type },
-                )}
-              />
-            </span>
-            <div className="file-structure fade-in panel-group">
-              <div className="panel panel-default">
-                <div className="panel-collapse" aria-expanded="false">
-                  <div className="panel-body">
-                    <FileTree
-                      tree={run[type]}
-                      treeId={run._id}
-                      editable={false}
-                      getFileDownloadTicket={actions.getResultDownloadTicket.bind(
-                        this,
-                        run.snapshotId,
-                        run._id,
-                      )}
-                      displayFile={this.props.displayFile.bind(
-                        this,
-                        run.snapshotId,
-                        run._id,
-                      )}
-                      toggleFolder={this.props.toggleFolder}
-                    />
-                  </div>
-                </div>
-              </div>
-            </div>
-          </Panel>
-        </Accordion>
-      )
-    }
   }
 
   _parameters(run) {
