+++ app/src/scripts/front-page/front-page.pipelines.jsx
@@ -7,7 +7,7 @@ import FileDisplay from '../dataset/dataset.file-display.jsx'
 import FPStore from './front-page.store.js'
 import FPActions from './front-page.actions.js'
 import Select from 'react-select'
-import Run from '../dataset/dataset.jobs.run.jsx'
+import Run from '../dataset/run'
 import { Link } from 'react-router-dom'
 import { Panel } from 'react-bootstrap'
 import pluralize from 'pluralize'
